Det vart så mykje mapping i Brevoppretter no at det slukte heile klassa, så trekte akkurat den biten ut til ei eiga klasse med kun det ansvaret
