+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/Brevoppretter.kt
@@ -1,31 +1,21 @@
 package no.nav.etterlatte.brev
 
-import kotlinx.coroutines.async
-import kotlinx.coroutines.coroutineScope
 import no.nav.etterlatte.brev.adresse.AdresseService
 import no.nav.etterlatte.brev.behandling.PersonerISak
-import no.nav.etterlatte.brev.behandling.avsender
-import no.nav.etterlatte.brev.brevbaker.BrevbakerRequest
-import no.nav.etterlatte.brev.brevbaker.BrevbakerService
 import no.nav.etterlatte.brev.db.BrevRepository
-import no.nav.etterlatte.brev.hentinformasjon.BrevdataFacade
 import no.nav.etterlatte.brev.hentinformasjon.behandling.BehandlingService
 import no.nav.etterlatte.brev.model.Adresse
 import no.nav.etterlatte.brev.model.Brev
 import no.nav.etterlatte.brev.model.BrevDataRedigerbar
-import no.nav.etterlatte.brev.model.BrevInnhold
-import no.nav.etterlatte.brev.model.BrevInnholdVedlegg
 import no.nav.etterlatte.brev.model.BrevProsessType
 import no.nav.etterlatte.brev.model.BrevkodeRequest
 import no.nav.etterlatte.brev.model.Mottaker
 import no.nav.etterlatte.brev.model.OpprettNyttBrev
-import no.nav.etterlatte.brev.model.Spraak
 import no.nav.etterlatte.libs.common.behandling.SakType
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
 import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
 import no.nav.etterlatte.libs.common.person.UkjentVergemaal
 import no.nav.etterlatte.libs.common.person.Vergemaal
-import no.nav.etterlatte.libs.common.retryOgPakkUt
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
 import no.nav.etterlatte.libs.ktor.token.BrukerTokenInfo
 import no.nav.etterlatte.libs.ktor.token.Saksbehandler
@@ -34,10 +24,8 @@ import java.util.UUID
 class Brevoppretter(
     private val adresseService: AdresseService,
     private val db: BrevRepository,
-    private val brevdataFacade: BrevdataFacade,
     private val behandlingService: BehandlingService,
-    private val brevbaker: BrevbakerService,
-    private val redigerbartVedleggHenter: RedigerbartVedleggHenter,
+    private val datainnhenterForBrevoppretting: DatainnhenterForBrevoppretting,
 ) {
     suspend fun opprettVedtaksbrev(
         sakId: Long,
@@ -77,7 +65,7 @@ class Brevoppretter(
         brevDataMapping: suspend (RedigerbarTekstRequest) -> BrevDataRedigerbar,
     ): Pair<Brev, String> =
         with(
-            hentInnData(
+            datainnhenterForBrevoppretting.hentInnData(
                 sakId,
                 behandlingId,
                 bruker,
@@ -111,7 +99,7 @@ class Brevoppretter(
         val spraak = db.hentBrevInnhold(brevId)?.spraak
 
         with(
-            hentInnData(
+            datainnhenterForBrevoppretting.hentInnData(
                 sakId,
                 behandlingId,
                 bruker,
@@ -131,89 +119,6 @@ class Brevoppretter(
             return BrevService.BrevPayload(
                 innhold.payload ?: db.hentBrevPayload(brevId),
                 innholdVedlegg ?: db.hentBrevPayloadVedlegg(brevId),
-            )
-        }
-    }
-
-    private suspend fun hentInnData(
-        sakId: Long,
-        behandlingId: UUID?,
-        bruker: BrukerTokenInfo,
-        brevKode: (b: BrevkodeRequest) -> EtterlatteBrevKode,
-        brevDataMapping: suspend (RedigerbarTekstRequest) -> BrevDataRedigerbar,
-        overstyrSpraak: Spraak? = null,
-    ): OpprettBrevRequest {
-        val generellBrevData =
-            retryOgPakkUt { brevdataFacade.hentGenerellBrevData(sakId, behandlingId, overstyrSpraak, bruker) }
-
-        val brevkodeRequest =
-            BrevkodeRequest(
-                generellBrevData.loependeIPesys(),
-                generellBrevData.erForeldreloes(),
-                generellBrevData.sak.sakType,
-                generellBrevData.forenkletVedtak?.type,
-            )
-
-        val kode = brevKode(brevkodeRequest)
-        val tittel = kode.tittel ?: (generellBrevData.vedtakstype()?.let { "Vedtak om $it" } ?: "Tittel mangler")
-        return coroutineScope {
-            val redigerbarTekstRequest =
-                RedigerbarTekstRequest(
-                    brukerTokenInfo = bruker,
-                    soekerOgEventuellVerge = generellBrevData.personerISak.soekerOgEventuellVerge(),
-                    sakType = generellBrevData.sak.sakType,
-                    forenkletVedtak = generellBrevData.forenkletVedtak,
-                    utlandstilknytningType = generellBrevData.utlandstilknytning?.type,
-                    revurderingaarsak = generellBrevData.revurderingsaarsak,
-                    behandlingId = behandlingId,
-                    erForeldreloes = generellBrevData.erForeldreloes(),
-                    loependeIPesys = generellBrevData.loependeIPesys(),
-                    systemkilde = generellBrevData.systemkilde,
-                    avdoede = generellBrevData.personerISak.avdoede,
-                )
-            val brevData: BrevDataRedigerbar = brevDataMapping(redigerbarTekstRequest)
-            val innhold =
-                async {
-                    brevbaker.hentRedigerbarTekstFraBrevbakeren(
-                        BrevbakerRequest.fra(
-                            brevKode = kode,
-                            brevData = brevData,
-                            avsender =
-                                adresseService.hentAvsender(
-                                    avsender(bruker, generellBrevData.forenkletVedtak, generellBrevData.sak.enhet),
-                                ),
-                            soekerOgEventuellVerge = generellBrevData.personerISak.soekerOgEventuellVerge(),
-                            sakId = sakId,
-                            spraak = generellBrevData.spraak,
-                            sakType = generellBrevData.sak.sakType,
-                        ),
-                    )
-                }
-
-            val innholdVedlegg =
-                async {
-                    redigerbartVedleggHenter.hentInitiellPayloadVedlegg(
-                        bruker,
-                        kode.brevtype,
-                        generellBrevData.sak.sakType,
-                        generellBrevData.behandlingId,
-                        generellBrevData.revurderingsaarsak,
-                        generellBrevData.forenkletVedtak?.type,
-                        generellBrevData.sak.id,
-                        generellBrevData.spraak,
-                        generellBrevData.forenkletVedtak,
-                        generellBrevData.sak.enhet,
-                        generellBrevData.personerISak.soekerOgEventuellVerge(),
-                    )
-                }
-
-            OpprettBrevRequest(
-                innhold = BrevInnhold(tittel, generellBrevData.spraak, innhold.await()),
-                innholdVedlegg = innholdVedlegg.await(),
-                soekerFnr = generellBrevData.personerISak.soeker.fnr.value,
-                sakType = generellBrevData.sak.sakType,
-                enhet = generellBrevData.sak.enhet,
-                personerISak = generellBrevData.personerISak,
             )
         }
     }
@@ -247,15 +152,6 @@ class Brevoppretter(
             adresse = Adresse(adresseType = "", landkode = "", land = ""),
         )
 }
-
-private data class OpprettBrevRequest(
-    val innhold: BrevInnhold,
-    val innholdVedlegg: List<BrevInnholdVedlegg>?,
-    val soekerFnr: String,
-    val sakType: SakType,
-    val enhet: String,
-    val personerISak: PersonerISak,
-)
 
 class KanIkkeOppretteVedtaksbrev(
     behandlingId: UUID,
