+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/brev/varselbrev/VarselbrevTest.kt
@@ -6,6 +6,7 @@ import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
 import no.nav.etterlatte.brev.Brevoppretter
 import no.nav.etterlatte.brev.DatabaseExtension
+import no.nav.etterlatte.brev.DatainnhenterForBrevoppretting
 import no.nav.etterlatte.brev.PDFGenerator
 import no.nav.etterlatte.brev.RedigerbartVedleggHenter
 import no.nav.etterlatte.brev.adresse.AdresseService
@@ -112,14 +113,14 @@ class VarselbrevTest(
                 } returns listOf()
             }
         val behandlingService = mockk<BehandlingService>().also { coEvery { it.hentSak(sak.id, any()) } returns sak }
+        val datainnhenterForBrevoppretting =
+            DatainnhenterForBrevoppretting(brevdataFacade, brevbaker, adresseService, redigerbartVedleggHenter)
         val brevoppretter =
             Brevoppretter(
                 adresseService,
                 brevRepository,
-                brevdataFacade,
                 behandlingService,
-                brevbaker,
-                redigerbartVedleggHenter,
+                datainnhenterForBrevoppretting,
             )
         val pdfGenerator = mockk<PDFGenerator>()
         service = VarselbrevService(brevRepository, brevoppretter, behandlingService, pdfGenerator, mockk())
